chore(jans-cedarling): extract `BootstrapConfig` to module bootstrap_config
