+++ jans-cedarling/cedarling/src/log/log_strategy.rs
@@ -9,8 +9,8 @@ use super::interface::{LogStorage, LogWriter};
 use super::memory_logger::MemoryLogger;
 use super::nop_logger::NopLogger;
 use super::stdout_logger::StdOutLogger;
-use crate::models::log_config::LogConfig;
-use crate::models::log_config::LogTypeConfig;
+use crate::bootstrap_config::log_config::LogConfig;
+use crate::bootstrap_config::log_config::LogTypeConfig;
 use crate::models::log_entry::LogEntry;
 
 /// LogStrategy implements strategy pattern for logging.
